Added some safety values for draw boundaries, fixed mistake in initial player spawn location saving as int instead of double, also fixed player always to spawn in middle of the tile

+++ inc/cube.h
@@ -6,7 +6,7 @@
 /*   By: skwon2 <skwon2@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 17:11:37 by jmakkone          #+#    #+#             */
-/*   Updated: 2024/11/15 15:08:41 by jmakkone         ###   ########.fr       */
+/*   Updated: 2024/11/16 19:29:13 by jmakkone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,8 +95,8 @@ typedef struct s_map
 	char	**map_arr;
 	double	scale_x;
 	double	scale_y;
-	int		spawn_location_x;
-	int		spawn_location_y;
+	double	spawn_location_x;
+	double	spawn_location_y;
 }	t_map;
 
 typedef struct s_minmap
